Extract `Workspace.Element` > `WorkspaceElement` within ProjectDescription
