+++ Sources/ProjectDescription/Project.swift
@@ -6,19 +6,16 @@ public class Project: Codable {
     public let name: String
     public let targets: [Target]
     public let settings: Settings?
-
-    public enum CodingKeys: String, CodingKey {
-        case name
-        case targets
-        case settings
-    }
+    public let additionalFiles: [WorkspaceElement]
 
     public init(name: String,
                 settings: Settings? = nil,
-                targets: [Target] = []) {
+                targets: [Target] = [],
+                additionalFiles: [WorkspaceElement] = []) {
         self.name = name
         self.targets = targets
         self.settings = settings
+        self.additionalFiles = additionalFiles
         dumpIfNeeded(self)
     }
 }
